refactor(test): move test package to dev.hypera
